refactor: effects module export
