+++ js/img-upload.js
@@ -1,4 +1,4 @@
-import './img-upload-effects.js';
+import {upScale, downScale, toggleEffect} from './img-upload-effects.js';
 import {openModal} from './modal.js';
 import {isValid} from './validate-form.js';
 
@@ -9,6 +9,9 @@ const imgHashtagsInput = document.querySelector('.text__hashtags');
 const imgDescriptionInput = document.querySelector('.text__description');
 const imgUploadButton = document.querySelector('.img-upload__submit');
 const previewImage = document.querySelector('.img-upload__preview > img');
+const imgScaleSmallerButton = document.querySelector('.scale__control--smaller');
+const imgScaleBiggerButton = document.querySelector('.scale__control--bigger');
+const effectsRadio = document.querySelector('.effects__list');
 
 imgUploadInput.addEventListener('change', () => {
   openModal(imgUploadOverlay, imgUploadInput, imgHashtagsInput, imgDescriptionInput);
@@ -17,6 +20,10 @@ imgUploadInput.addEventListener('change', () => {
 
   imgUploadForm.addEventListener('input', onFormFieldsInput);
   // TODO добавить снятие обработчиков при отправке формы
+
+  imgScaleSmallerButton.addEventListener('click', downScale);
+  imgScaleBiggerButton.addEventListener('click', upScale);
+  effectsRadio.addEventListener('click', toggleEffect);
 });
 
 function onFormFieldsInput() {
